refactor Timeline.tsx and server.ts
