+++ app/components/timeline/index.ts
@@ -1,0 +1,3 @@
+export * from "./utils";
+export { default as TimelineToolbar } from "./TimelineToolbar";
+export type { ToolMode, TransformMode } from "./TimelineToolbar";
